+++ src/__ray-template/index.ts
@@ -1,0 +1,2 @@
+export * from './valid/validAppRootPath'
+export * from './valid/validLocal'