+++ src/global-variable/variable.ts
@@ -11,10 +11,9 @@
 
 /**
  *
- * 存放全局响应式变量，脱离 pinia 使用的变量
- * 简单的全局变量，并不需要复杂的控制流程
- *
- * 使用方法
+ * @description
+ * 存放全局响应式变量，脱离 pinia 使用的变量。
+ * 简单的全局变量，并不需要复杂的控制流程。
  *
  * @example
  * 获取响应式变量
@@ -31,12 +30,11 @@ import type { AnyFC } from '@/types'
 
 /**
  *
- * 全局响应式变量
- * 不推荐直接访问、设置该 state。并且更不建议在组件中直接使用该 state
- * 请使用 `getVariable`、`getVariableToRefs`、`setVariable` 方法进行操作
+ * @description
+ * 全局响应式变量。
  *
- * 为什么会有该变量存在，是因为在有时候你希望在 vue 声明之后就调用该变量，但是又不想使用 pinia
- * 你可以使用该变量，但是请注意，该变量不会被 pinia 管理，所以你需要自己手动管理
+ * 不推荐直接访问、设置该 state不建议在组件中直接使用该 state，
+ * 请使用 `getVariable`、`getVariableToRefs`、`setVariable` 方法进行操作。
  */
 const variableState = reactive({
   globalSpinning: false, // 全局加载控制器
@@ -79,7 +77,7 @@ export function setVariable<T extends VariableStateKey, FC extends AnyFC>(
  * 返回目标值的响应式 ref
  *
  * @example
- * getVariableToRefs('globalSpinning') // 返回 ref<boolean>
+ * getVariableToRefs('globalSpinning') // 返回 Ref<boolean>
  */
 export function getVariableToRefs<K extends VariableStateKey>(key: K) {
   return readonly(toRef<VariableState, K>(variableState, key))